Update MagicNDIreceiver.cpp
15.12.21- Equalise version numbers for sender and receiver.
Update to the latest ofxNDI source with NDI SDK Version 5.0.10.1
Fix NDI version number length.
For user selection, set the selected sender name or the old one will be used.
This can happen if there is a name in the dialog at startup and it's not running.
Version 1.017
11.12.22 - Suppress warnings
12.12.22 - Add ReleaseNDIreceiver
Expand help and tooltips
Remove unused definitions and pbos
Remove unused Glee library
Allow for sender name change for the same index
to allow for Spout to NDI selecting another sender
Rebuild with revised ofxNDI
26.12.22 - Rebuild with revised ofxNDI
Version 1.018